+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- [#1867](https://github.com/teambit/bit/issues/1867) apply workspace overrides config on imported components
+- [#1892](https://github.com/teambit/bit/issues/1892) rename `--bare` flag of `bit list` to `--raw`
 - [#1863](https://github.com/teambit/bit/issues/1863) allow excluding components from `overrides` rules
 - [#1865](https://github.com/teambit/bit/issues/1865) allow adding `package.json` props via `overrides`
 - [#1837](https://github.com/teambit/bit/issues/1837) enable executing commands on remote components outside of bit-workspace